fix: moving component to a new folder and fixing imports
