+++ src/components/atoms/index.ts
@@ -3,5 +3,5 @@
 export { Button } from './Button/Button'
 export { Input } from './Input'
 export { Text, Heading, Price } from './Text'
-export { Badge } from './Badge'
+export { Badge } from './Badge/Badge'
 export { Icon } from './Icon'
